+++ pyodeint/__init__.py
@@ -5,7 +5,9 @@ Python binding for odeint from boost.
 
 from __future__ import absolute_import
 
-from ._odeint_numpy import adaptive, predefined, requires_jac, steppers
+import numpy as np
+
+from ._odeint import adaptive, predefined, requires_jac, steppers
 from ._util import _check_callable, _check_indexing
 
 from ._release import __version__
@@ -15,6 +17,12 @@ def get_include():
     from pkg_resources import resource_filename, Requirement
     return resource_filename(Requirement.parse(__name__),
                              '%s/include' % __name__)
+
+def _bs(kwargs):
+    # DEPRECATED accept 'bs' as short for 'bulirsh_stoer'
+    if kwargs.get('method', '-') == 'bs':
+        kwargs['method'] = 'bulirsch_stoer'
+    return kwargs
 
 
 def integrate_adaptive(rhs, jac, y0, x0, xend, dx0, atol, rtol,
@@ -64,7 +72,7 @@ def integrate_adaptive(rhs, jac, y0, x0, xend, dx0, atol, rtol,
     if check_indexing:
         _check_indexing(rhs, jac, x0, y0)
 
-    return adaptive(rhs, jac, y0, x0, xend, dx0, atol, rtol, **kwargs)
+    return adaptive(rhs, jac, np.asarray(y0, dtype=np.float64), x0, xend, dx0, atol, rtol, **_bs(kwargs))
 
 
 def integrate_predefined(rhs, jac, y0, xout, dx0, atol, rtol,
@@ -111,4 +119,5 @@ def integrate_predefined(rhs, jac, y0, xout, dx0, atol, rtol,
     if check_indexing:
         _check_indexing(rhs, jac, xout[0], y0)
 
-    return predefined(rhs, jac, y0, xout, dx0, atol, rtol, **kwargs)
+    return predefined(rhs, jac, np.asarray(y0, dtype=np.float64), np.asarray(xout, dtype=np.float64),
+                      dx0, atol, rtol, **_bs(kwargs))